Move ParallelCoordinate to tslib/react
